+++ neoscore/core/neoscore.py
@@ -210,7 +210,8 @@ def show(
     refresh_func: Optional[RefreshFunc] = None,
     display_page_geometry=True,
     auto_viewport_interaction_enabled=True,
-    window_size: Optional[tuple[int, int]] = None,
+    min_window_size: Optional[tuple[int, int]] = None,
+    max_window_size: Optional[tuple[int, int]] = None,
 ):
     """Display the score in an interactive GUI window.
 
@@ -224,7 +225,8 @@ def show(
         auto_viewport_interaction_enabled: Whether mouse and scrollbar viewport
             interaction is enabled. If false, scrollbars do not appear, mousewheel
             zooming is disabled, and click-and-drag view movement is disabled.
-        window_size: An optional ``(width, height)`` window size tuple.
+        min_window_size: An optional ``(width, height)`` minimum window size tuple.
+        max_window_size: An optional ``(width, height)`` maximum window size tuple.
     """
     global document
     global app_interface
@@ -235,7 +237,7 @@ def show(
     if refresh_func:
         set_refresh_func(refresh_func)
     app_interface.auto_viewport_interaction_enabled = auto_viewport_interaction_enabled
-    app_interface.show(window_size)
+    app_interface.show(min_window_size, max_window_size)
 
 
 def _clear_interfaces():